Create internal package and move all internal classes there
