+++ flink-connector/src/main/java/io/delta/flink/sink/DeltaSink.java
@@ -22,11 +22,11 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
-import io.delta.flink.sink.committables.DeltaCommittable;
-import io.delta.flink.sink.committables.DeltaGlobalCommittable;
-import io.delta.flink.sink.committer.DeltaGlobalCommitter;
-import io.delta.flink.sink.writer.DeltaWriter;
-import io.delta.flink.sink.writer.DeltaWriterBucketState;
+import io.delta.flink.sink.internal.committables.DeltaCommittable;
+import io.delta.flink.sink.internal.committables.DeltaGlobalCommittable;
+import io.delta.flink.sink.internal.committer.DeltaGlobalCommitter;
+import io.delta.flink.sink.internal.writer.DeltaWriter;
+import io.delta.flink.sink.internal.writer.DeltaWriterBucketState;
 import org.apache.flink.api.connector.sink.Committer;
 import org.apache.flink.api.connector.sink.GlobalCommitter;
 import org.apache.flink.api.connector.sink.Sink;
@@ -58,10 +58,10 @@ import io.delta.standalone.DeltaLog;
  * {@link org.apache.flink.connector.file.sink.FileSink}.
  * <p>
  * Next during the checkpoint phase files are "closed" (renamed) by the independent instances of
- * {@link io.delta.flink.sink.committer.DeltaCommitter} that behave very similar
+ * {@link io.delta.flink.sink.internal.committer.DeltaCommitter} that behave very similar
  * to {@link org.apache.flink.connector.file.sink.committer.FileCommitter}.
  * When all the parallel committers are done, then all the files are committed at once by
- * single-parallelism {@link io.delta.flink.sink.committer.DeltaGlobalCommitter}.
+ * single-parallelism {@link io.delta.flink.sink.internal.committer.DeltaGlobalCommitter}.
  * <p>
  * This {@link DeltaSink} sources many specific implementations from the
  * {@link org.apache.flink.connector.file.sink.FileSink} so for most of the low level behaviour one
