+++ flink-connector/src/main/java/io/delta/flink/sink/DeltaSinkBuilder.java
@@ -22,15 +22,15 @@ import java.io.IOException;
 import java.io.Serializable;
 import java.util.UUID;
 
-import io.delta.flink.sink.committables.DeltaCommittable;
-import io.delta.flink.sink.committables.DeltaCommittableSerializer;
-import io.delta.flink.sink.committables.DeltaGlobalCommittable;
-import io.delta.flink.sink.committables.DeltaGlobalCommittableSerializer;
-import io.delta.flink.sink.committer.DeltaCommitter;
-import io.delta.flink.sink.committer.DeltaGlobalCommitter;
-import io.delta.flink.sink.writer.DeltaWriter;
-import io.delta.flink.sink.writer.DeltaWriterBucketState;
-import io.delta.flink.sink.writer.DeltaWriterBucketStateSerializer;
+import io.delta.flink.sink.internal.committables.DeltaCommittable;
+import io.delta.flink.sink.internal.committables.DeltaCommittableSerializer;
+import io.delta.flink.sink.internal.committables.DeltaGlobalCommittable;
+import io.delta.flink.sink.internal.committables.DeltaGlobalCommittableSerializer;
+import io.delta.flink.sink.internal.committer.DeltaCommitter;
+import io.delta.flink.sink.internal.committer.DeltaGlobalCommitter;
+import io.delta.flink.sink.internal.writer.DeltaWriter;
+import io.delta.flink.sink.internal.writer.DeltaWriterBucketState;
+import io.delta.flink.sink.internal.writer.DeltaWriterBucketStateSerializer;
 import org.apache.flink.api.connector.sink.Sink;
 import org.apache.flink.core.fs.FileSystem;
 import org.apache.flink.core.fs.Path;
