+++ flink-connector/src/main/java/io/delta/flink/sink/internal/committables/DeltaCommittableSerializer.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.delta.flink.sink.committables;
+package io.delta.flink.sink.internal.committables;
 
 import java.io.IOException;
