+++ flink-connector/src/main/java/io/delta/flink/sink/internal/writer/DeltaWriter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.delta.flink.sink.writer;
+package io.delta.flink.sink.internal.writer;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -27,7 +27,7 @@ import java.util.Map;
 import javax.annotation.Nullable;
 
 import io.delta.flink.sink.DeltaTablePartitionAssigner;
-import io.delta.flink.sink.committables.DeltaCommittable;
+import io.delta.flink.sink.internal.committables.DeltaCommittable;
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.connector.sink.Sink;
@@ -264,8 +264,8 @@ public class DeltaWriter<IN>
 
     /**
      * This method prepares committables objects that will be passed to
-     * {@link io.delta.flink.sink.committer.DeltaCommitter} and
-     * {@link io.delta.flink.sink.committer.DeltaGlobalCommitter} to finalize the
+     * {@link io.delta.flink.sink.internal.committer.DeltaCommitter} and
+     * {@link io.delta.flink.sink.internal.committer.DeltaGlobalCommitter} to finalize the
      * checkpoint interval and commit written files.
      *
      * @implNote This method behaves in the same way as
