+++ flink-connector/src/test/java/io/delta/flink/sink/internal/TestSchemaConverter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.delta.flink.sink;
+package io.delta.flink.sink.internal;
 
 import java.util.ArrayList;
 import java.util.Arrays;
