+++ flink-connector/src/test/java/io/delta/flink/sink/internal/committer/DeltaGlobalCommitterTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.delta.flink.sink.committer;
+package io.delta.flink.sink.internal.committer;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -28,10 +28,10 @@ import java.util.Random;
 import java.util.UUID;
 import java.util.stream.Stream;
 
-import io.delta.flink.sink.SchemaConverter;
-import io.delta.flink.sink.committables.DeltaCommittable;
-import io.delta.flink.sink.committables.DeltaGlobalCommittable;
-import io.delta.flink.sink.committables.DeltaGlobalCommittableSerializer;
+import io.delta.flink.sink.internal.SchemaConverter;
+import io.delta.flink.sink.internal.committables.DeltaCommittable;
+import io.delta.flink.sink.internal.committables.DeltaGlobalCommittable;
+import io.delta.flink.sink.internal.committables.DeltaGlobalCommittableSerializer;
 import io.delta.flink.sink.utils.DeltaSinkTestUtils;
 import org.apache.flink.connector.file.sink.utils.FileSinkTestUtils;
 import org.apache.flink.core.fs.Path;
